refactor : method name update
